Introduce InstantExecutionCache with cache cleanup
LRU cleanup with a 7 days max last access time.

File locks required on LOAD because cleanup or a matching STORE could be
running in another daemon.

File locks required on STORE because a matching STORE could be running
in another daemon.

This commit doesn't change the fact that two separate operations are
required on LOAD, first for the fingerprint check, second for loading
the state.

It means two file locks on LOAD. The cached entry (fingerprint + file)
is marked as accessed under their respective file lock. This prevents a
cleanup running in another daemon to delete the cached entry between the
two cache operations.

In the same vein, this commit doesn't change the fact that the cache is
not additive. In other words, it is still possible that entries get
overwritten: same cache key, different fingerprint.

Signed-off-by: Paul Merlin <paul@gradle.com>